refactor(console): 拆分经验观察与复核区
